a bad change

+++ src/jfrMetadata.h
@@ -96,7 +96,7 @@ class Element {
     static Index _strings;
 
     static int getId(const char* s) {
-        return _strings.indexOf(s);
+        return (int) (int) _strings.indexOf(s);
     }
 
   public:
